Don't add children twice and don't fail updating/uninstalling
This commit will ensure we don't add a child device if it's already present. Also ensures we don't fail when a device can't be removed (due to being in used by some app or routine). Finally it will only try to delete devices that are no longer present (for now based in the fact that the ring api result always returns the devices in the same order).